+++ src/core/NamespaceStore.js
@@ -2,28 +2,31 @@
 export default class NamespaceStore {
 
   constructor() {
-    this.static = {};
-    this.instance = {};
+    this.store = {};
   }
 
-  has(name, isInstance) {
-    return this.getStore(isInstance).hasOwnProperty(name);
+  has(name, member, isInstance) {
+    const key = this.getKey(name, member, isInstance);
+    return this.store.hasOwnProperty(key);
   }
 
-  get(name, isInstance) {
-    return this.getStore(isInstance)[name];
+  get(name, member, isInstance) {
+    const key = this.getKey(name, member, isInstance);
+    return this.store[key];
   }
 
-  set(name, val, isInstance) {
-    this.getStore(isInstance)[name] = val;
+  set(name, member, val, isInstance) {
+    const key = this.getKey(name, member, isInstance);
+    this.store[key] = val;
   }
 
-  remove(name, isInstance) {
-    delete this.getStore(isInstance)[name];
+  remove(name, member, isInstance) {
+    const key = this.getKey(name, member, isInstance);
+    delete this.store[key];
   }
 
-  getStore(isInstance) {
-    return isInstance ? this.instance : this.static;
+  getKey(name, member, isInstance) {
+    return name + (isInstance ? '#' : '.') + member;
   }
 
 }